Domain

This folder contains structures that are used to encode different kinds of concrete and abstract domains.

Concrete continuous domains

Dionysos.Domain.ContinuousUnboundedDomain — Type

ContinuousUnboundedDomain{N,T}

Struct for a basic unbounded continuous domain

Dionysos.Domain.ContinuousBoundedDomain — Type

ContinuousBoundedDomain{N,T,B}

Struct for a basic bounded continuous domain

Dionysos.Domain.ContinuousBoundedEllipsoidDomain — Type

ContinuousBoundedEllipsoidDomain{N,T,S<:Grid{N,T}}

Struct for a basic bounded continuous domain formed by a finite number of ellipsoids

Dionysos.Domain.RectangularObstacles — Type

RectangularObstacles{VT} <: AbstractSet{VT}

Struct for a rectangular domain with rectangular obstacles

Abstract domains

Dionysos.Domain.DomainList — Type

DomainList{N,T,S<:Grid{N,T}}

Struct for a basic domain based on a Grid

Dionysos.Domain.GeneralDomainList — Type

GeneralDomainList{N,E<:AbstractSet{NTuple{N,Int}},T,S<:Grid{N,T},F} <: DomainType{N,T}

Struct for a rectangular domain with rectangular obstacles

Dionysos.Domain.GridRectangular — Type

GridRectangular{N,T} <: Grid{N,T}

Uniform grid on rectangular space rect, centered at orig and with steps set by the vector h.

Dionysos.Domain.GridFree — Type

GridFree{N,T} <: Grid{N,T}

Uniform grid on unbounded space, centered at orig and with steps set by the vector h.

Dionysos.Domain.GridEllipsoidalRectangular — Type

GridEllipsoidalRectangular{N,T} <: Grid{N,T}

Uniform grid on rectangular space rect, centered at orig and with steps set by the vector h. Cells are (possibly overlapping) ellipsoids defined at each grid point c as (x-c)'P(x-c) ≤ 1