Mapping

This folder contains mappings to make the link between the concrete domain and the abstract domain, and vice-versa.

Dionysos.Mapping.MappingContinuousGrid — Type

MappingContinuousGrid(C::DO.ContinuousDomain, D::DomainList{N,T,S<:DO.Grid{N,T}}

Mapping containing methods to convert elements in a ContinuousDomain to a grid-based DomainList and vice-versa.

Dionysos.Mapping.MappingContinuousEllipsoid — Type

MappingContinuousEllipsoid{C<:DO.ContinuousDomain{N,T}, CE<:DO.ContinuousBoundedEllipsoidDomain{N,T,B,E}}

Mapping containing methods to convert elements in a ContinuousDomain to a ContinuousBoundedEllipsoidDomain and vice-versa.