Optim

This folder contains all the different (abstraction-based or not) solvers that can be used. Note that all the solvers are defined using the MathOptInterface framework: for each solver, we define a subtype of AbstractOptimizer and implement the Optimize! function.

Abstraction-based solvers

Dionysos.Optim.Abstraction.UniformGridAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

A solver implementing the classical abstraction method (e.g., used in SCOTS), where the entire domain is partitioned into hyper-rectangular cells, independent of the specific control task. 

The optimizer is structured into modular sub-solvers, each dedicated to a specific problem type. It ensures that abstraction is computed before solving the control problem, maintaining modularity and extendability.

Dionysos.Optim.Abstraction.EllipsoidsAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Abstraction-based solver for which the domain is covered with ellipsoidal cells, independently of the control task.

Dionysos.Optim.Abstraction.HierarchicalAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Abstraction-based solver for which the domain is initially partioned into coarse hyper-rectangular cells, which are iteratively locally smartly refined with respect to the control task.

Dionysos.Optim.Abstraction.LazyAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Abstraction-based solver for which the hyper-rectangular abstraction and the controller are co-designed to reduce the computation cost of the abstraction.

Dionysos.Optim.Abstraction.LazyEllipsoidsAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Abstraction-based solver using the lazy abstraction method with ellipsoidal cells.

Other solvers

Dionysos.Optim.BemporadMorari.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Bemporad Morari solver: Optimal control of hybrid systems via a predictive control scheme using mixed integer quadratic programming (MIQP) online optimization procedures.

Dionysos.Optim.BranchAndBound.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Branch and bound solver: Optimal control of hybrid systems via a predictive control scheme combining a branch and bound algorithm that can refine Q-functions using Lagrangian duality.

Solvers details

Uniform Grid Abstraction

Dionysos.Optim.Abstraction.UniformGridAbstraction.solve_concrete_problem — Function

solve_concrete_problem(abstract_system, abstract_controller)

Solves the concrete problem based on the abstract_system and abstract_controller.

Dionysos.Optim.Abstraction.UniformGridAbstraction._validate_model — Function

validate_model!(model::OptimizerEmptyProblem, required_fields::Vector{Symbol})

Ensures the specified required_fields are set in the model. Throws an error if any field is missing.

Dionysos.Optim.Abstraction.UniformGridAbstraction._validate_continuous_model — Function

validate_continuous_model!(model::OptimizerEmptyProblem)

Validates that the OptimizerEmptyProblem model is correctly configured for continuous systems. Throws an error if required fields like time_step or jacobian_bound are missing or invalid.

Dionysos.Optim.Abstraction.UniformGridAbstraction._validate_discrete_model — Function

validate_discrete_model!(model::OptimizerEmptyProblem)

Validates that the OptimizerEmptyProblem model is correctly configured for discrete systems. Throws an error if required fields like sys_inv_map or growthbound_map are missing or invalid.

Dionysos.Optim.Abstraction.UniformGridAbstraction._maybe_discretized_system — Function

_maybe_discretized_system(concrete_system, model, noise)

Returns the discretized system based on the concrete_system and model.

Dionysos.Optim.Abstraction.UniformGridAbstraction._discrete_system — Function

discretized_system(concrete_system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem, model, noise)

Returns the discretized system based on the concrete_system and model.

Dionysos.Optim.Abstraction.UniformGridAbstraction._discretize_continuous_system — Function

_discretize_continuous_system(concrete_system::MathematicalSystems.ConstrainedBlackBoxControlContinuousSystem, model, noise)

Returns the discretized system based on the concrete_system and model.

Dionysos.Optim.Abstraction.UniformGridAbstraction._get_domain_list — Function

_get_domain_list(model, variables, grid)

Returns a Dionysos.Domain.DomainList based on the variables and grid.

Dionysos.Optim.Abstraction.UniformGridAbstraction.compute_largest_controllable_set — Function

compute_largest_controllable_set(abstract_system, target_set; initial_set)

Computes the largest controllable set and generates an abstract controller.

Dionysos.Optim.Abstraction.UniformGridAbstraction.compute_largest_invariant_set — Function

compute_largest_invariant_set(abstract_system, setlist)

Computes the largest controllable invariant set within safelist.