Optim

This folder contains all the different (abstraction-based or not) solvers that can be used. Note that all the solvers are defined using the MathOptInterface framework: for each solver, we define a subtype of AbstractOptimizer and implement the Optimize! function.

Abstraction-based solvers

Dionysos.Optim.Abstraction.UniformGridAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Solver based on the classical abstraction method (used for instance in SCOTS) for which the whole domain is partioned into hyper-rectangular cells, independently of the control task.

Dionysos.Optim.Abstraction.EllipsoidsAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Abstraction-based solver for which the domain is covered with ellipsoidal cells, independently of the control task.

Dionysos.Optim.Abstraction.HierarchicalAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Abstraction-based solver for which the domain is initially partioned into coarse hyper-rectangular cells, which are iteratively locally smartly refined with respect to the control task.

Dionysos.Optim.Abstraction.LazyAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Abstraction-based solver for which the hyper-rectangular abstraction and the controller are co-designed to reduce the computation cost of the abstraction.

Dionysos.Optim.Abstraction.LazyEllipsoidsAbstraction.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Abstraction-based solver using the lazy abstraction method with ellipsoidal cells.

Other solvers

Dionysos.Optim.BemporadMorari.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Bemporad Morari solver: Optimal control of hybrid systems via a predictive control scheme using mixed integer quadratic programming (MIQP) online optimization procedures.

Dionysos.Optim.BranchAndBound.Optimizer — Type

Optimizer{T} <: MOI.AbstractOptimizer

Branch and bound solver: Optimal control of hybrid systems via a predictive control scheme combining a branch and bound algorithm that can refine Q-functions using Lagrangian duality.