Problem

This folder contains structures that are used to encode which kind of problem you want to solve. Each problem is encoded as a ProblemType.

Dionysos.Problem.ProblemType — Type

The structure

ProblemType

is the abstract type which defines a problem.

So far, two types of problems have been considered: the reach-avoid optimal control problems and the safety control problems.

Dionysos.Problem.OptimalControlProblem — Type

The structure

OptimalControlProblem{S, XI, XT, XC, TC, T}

encodes an optimal control problem where

S is the system,

XI is the initial set,

XT is the target set,

XC is the state cost,

TC is transistion cost and

T is the number of allowed time steps

Dionysos.Problem.SafetyProblem — Type

The structure

SafetyProblem{S, XI, XS, T}

encodes a safety problem where

S is the system,

XI is the initial set,

XS is the safe set and

T is the number of allowed time steps