Symbolic

This folder contains the data structures needed to encode the different abstractions.

Dionysos.Symbolic.SymbolicModel — Type

SymbolicModel{N, M}

is the abtract type which defines a symbolic model.

Dionysos.Symbolic.SymbolicModelList — Type

SymbolicModelList{N, M, S1 <: DO.DomainType{N}, S2 <: DO.DomainType{M}, A} <: SymbolicModel{N, M}

is one implementation of the SymbolicModel type for classical abstraction-based methods, i.e. when the whole domain is partitioned/covered.

Dionysos.Symbolic.LazySymbolicModel — Type

LazySymbolicModel{N, M, S1 <: DO.DomainType{N}, S2 <: DO.DomainType{M}, A} <: SymbolicModel{N, M}

is one implementation of the SymbolicModel type for the lazy abstraction-based methods, i.e. when a subset of the domain is partitioned/covered.

Dionysos.Symbolic.compute_symmodel_from_hybridcontrolsystem! — Function

compute_symmodel_from_hybridcontrolsystem!(symmodel::SymbolicModel{N}, transitionCost::AbstractDict, transitionCont::AbstractDict,
hybridsys::AbstractHybridSystem, W, L, U, opt_sdp, opt_qp)

Builds an abstraction symmodel where the transitions have costs given in transitionCost and are parameterized by affine-feedback controllers in transitionCont. The concrete system is hybridsys and W, L and U are defined as in _has_transition. An SDP optimizer opt_sdp and a QP optimizer opt_qp must be provided as JuMP optimizers.