System

This folder contains different ways to define systems, for instance to encode a controller.

Control system

Dionysos.System.ControlSystemGrowth — Type

ControlSystemGrowth{N, T, F1 <: Function, F2 <: Function, F3 <: Function} <: ControlSystem{N, T}

TO ADD

Dionysos.System.ControlSystemLinearized — Type

ControlSystemLinearized{N, T, F1 <: Function, F2 <: Function, F3 <: Function, F4 <: Function, } <: ControlSystem{N, T}

TO ADD

Dionysos.System.SimpleSystem — Type

SimpleSystem{N, T, F <: Function, F2} <: ControlSystem{N, T}

TO ADD

Dionysos.System.EllipsoidalAffineApproximatedSystem — Type

EllipsoidalAffineApproximatedSystem{}

TO ADD

Dionysos.System.AffineApproximationDiscreteSystem — Type

AffineApproximationDiscreteSystem

TO ADD

Dionysos.System.SymbolicSystem — Type

SymbolicSystem{}

TO ADD

Controller

Dionysos.System.ConstantController — Type

ConstantController{T, VT}

encodes a constant state-dependent controller of the κ(x) = c

Dionysos.System.AffineController — Type

AffineController{T, MT, VT1, VT2}

encodes an affine state-dependent controller of the κ(x) = K*(x-c)+ℓ

SHOULD BE REMOVED

Dionysos.Control.DiscreteTrajectory — Type

DiscreteTrajectory{Q, TT}

q_0 is the starting mode and transitions is a sequence of discrete transitions in the system

Dionysos.Control.ContinuousTrajectory — Type

ContinuousTrajectory{T, XVT<:AbstractVector{T}, UVT<:AbstractVector{T}}

x is a sequence of points in the state space and u is a sequence of points in the input space