Temp

Temporary API Reference file

Dionysos.Utils.expand — Function

Yield the nodes reachable from this node.

Dionysos.Utils.BDD.IntTupleSet — Type

mutable struct IntTupleSet <: AbstractSet{NTuple{N,Int} where N}

Same as Base.Set{NTuple{N,Int} where N} but with CUDD.

Dionysos.Utils.path_cost — Function

Return the cost of a solution path that arrives at state2 from state1 via action, assuming cost c to get up to state1. If the problem is such that the path doesn't matter, this function will only look at state2. If the path does matter, it will consider c and maybe state1 and action. The default method costs 1 for every step in the path.

Dionysos.Symbolic.compute_symmodel_from_hybridcontrolsystem! — Function

compute_symmodel_from_hybridcontrolsystem!(symmodel::SymbolicModel{N}, transitionCost::AbstractDict, transitionKappa::AbstractDict,
hybridsys::AbstractHybridSystem, W, L, U, opt_sdp, opt_qp)

Builds an abstraction symmodel where the transitions have costs given in transitionCost and are parameterized by affine-feedback controllers in transitionKappa. The concrete system is hybridsys and W, L and U are defined as in _has_transition. An SDP optimizer opt_sdp and a QP optimizer opt_qp must be provided as JuMP optimizers.

Dionysos.Utils.breadth_first_graph_search — Function

Search the shallowest nodes in the search tree first.

Dionysos.Utils.depth_first_graph_search — Function

Search the deepest nodes in the search tree first.

Dionysos.Utils.best_first_graph_search — Function

Search the nodes with the lowest f scores first. You specify the function f(node) that you want to minimize; for example, if f is a heuristic estimate to the goal, then we have greedy best first search; if f is node.depth then we have depth-first search.

Dionysos.Utils.path — Function

Create a list of nodes from the root to this node.

Dionysos.Utils.AbstractQueue — Type

AbstractQueue is an abstract type. There are three types: MyStack(): A Last In First Out Queue. FIFOQueue(): A First In First Out Queue. MyPriorityQueue(f,ext): Queue where items are sorted by f, (default <). Each type supports the following methods and functions: append!(q,item) – add an item to the queue extend!(q,items) – equivalent to: for item in items: append(q,item) pop!(q) – return the top item from the queue length(q) – number of items in q

Dionysos.Symbolic._compute_base_cell — Function

_compute_base_cell(r::SVector{S})

Computes a polyhedron containing the base hyperrectangular cell, centered at the origin and with the i-th side lenght given by 2*r[i].

Dionysos.Utils.MyStack — Type

Return an empty list, suitable as a Last-In-First-Out Queue.

Dionysos.Control.DiscreteTrajectory — Type

DiscreteTrajectory{Q, TT}

q_0 is the starting mode and transitions is a sequence of discrete transitions in the system

Dionysos.Utils.tree_search — Function

Search through the successors of a problem to find a goal. The argument fringe should be an empty queue. Don't worry about repeated paths to a state.

Dionysos.Symbolic._get_min_bounding_box — Function

_get_min_bounding_box(P, optimizer)

Finds the minimum bounding box containing the ellipsoid {x'Px < 1}.

Dionysos.Utils.goal_test — Function

Return True if the state is a goal. The default method compares the state to P.goal, as specified in the constructor. Implement this method if checking against a single goal is not enough.

Dionysos.Utils.astar_graph_search — Function

A* search is best-first graph search with f(n) = g(n)+h(n). You need to specify the h function when you call astar_search.

Dionysos.Utils.astar_tree_search — Function

A* search is best-first graph search with f(n) = g(n)+h(n). You need to specify the h function when you call astar_search.

Dionysos.Utils.BDD.BitSet — Type

mutable struct BitSet <: AbstractSet{Int}

Same as Base.BitSet but with CUDD.

Dionysos.Symbolic._provide_P — Function

_provide_P(subsys::HybridSystems.ConstrainedAffineControlDiscreteSystem, optimizer)

If subsys is a stabilizable system, finds the matrix P and the state-feedback gain K that satisfy the discrete-time Lyapunov inequality (A+BK)'P(A+BK)-P < 0. The condition number of P is minimized. optimizer must be a JuMP SDP optimizer.

Dionysos.Utils.breadth_first_tree_search — Function

Search the shallowest nodes in the search tree first.

Dionysos.Utils.graph_search — Function

Search through the successors of a problem to find a goal. The argument fringe should be an empty queue. If two paths reach a state, only use the best one.

Dionysos.Symbolic.ellipsoid_vol — Function

ellipsoid_vol(P,r)

Calculates the n-volume of the n-ellipsoid defined as {x'Px < r}.

Dionysos.Symbolic._has_transition — Function

_has_transition(subsys::Union{HybridSystems.ConstrainedAffineControlDiscreteSystem,HybridSystems.HybridSystems.ConstrainedAffineControlMap},#
P,c,Pp,cp,W,L,U, optimizer)

Verifies whether a controller u(x)=K(x-c)+ell exists for subsys satisfying input requirements defined by U ans performs a sucessful transitions from a starting set Bs = {(x-c)'P(x-c) ≤ 1} to the final set Bs = {(x-c)'P(x-c) ≤ 1}. A tight upper bound on the transition cost c(x,u) is minimized where c(x,u) = |L*[x; u; 1]|^2, from the parameter L and each columm of the matrix W defines a vertex of the polytope from which additive disturbance are drawn.

The input restrictions are defined by the list U as |U[i]*u| ≤ 1. optimizer must be a JuMP SDP optimizer (e.g., Mosek, SDPA, COSMO, ...).

Note

This implements the optimization problem presented in Corollary 1 of the following paper https://arxiv.org/pdf/2204.00315.pdf

Dionysos.Utils.best_first_tree_search — Function

Search the nodes with the lowest f scores first. You specify the function f(node) that you want to minimize; for example, if f is a heuristic estimate to the goal, then we have greedy best first search; if f is node.depth then we have depth-first search.

Dionysos.Utils.successor — Function

Given a state, return a sequence of (action, state) pairs reachable from this state. If there are many successors, consider an iterator that yields the successors one at a time, rather than building them all at once.

Dionysos.Domain.DomainList — Type

DomainList{N,T,S<:Grid{N,T}}

Struct for a basic domain

Dionysos.Utils.Node — Type

A node in a search tree. Contains a pointer to the parent (the node that this is a successor of) and to the actual state for this node. Note that if a state is arrived at by two paths, then there are two nodes with the same state. Also includes the action that got us to this state, and the total pathcost (also known as g) to reach the node. Other functions may add an f and h value; see bestfirstgraphsearch and astar_search for an explanation of how the f and h values are handled.

Dionysos.Utils.depth_first_tree_search — Function

Search the deepest nodes in the search tree first.

Dionysos.Utils.MyPriorityQueue — Type

A queue in which the minimum (or maximum) element (as determined by f) is returned first. Keys of type T and priorities of type V.

Dionysos.Utils.FIFOQueue — Type

A First-In-First-Out Queue.

Dionysos.Control.ContinuousTrajectory — Type

ContinuousTrajectory{T, XVT<:AbstractVector{T}, UVT<:AbstractVector{T}}

x is a sequence of points in the state space and u is a sequence of points in the input space

Dionysos.Utils.SearchProblem — Type

SearchProblem

Fields

- `initial`  -- initial state of type `S` or a list of initial state of type `S`.
- `goal`     -- possibly a goal state of type 'Union{Nothing,S}''.

Example

struct Problem{S} <: SearchProblem{S}
    initial::Union{S,Vector{S}}
    goal::Union{Nothing,S}
end
The constructor specifies the initial state, and possibly a goal
state, if there is a unique goal.
function Problem(initial::S; goal=nothing) where S
    return Problem{S}(initial,goal)
end

Dionysos.Utils.BranchAndBound.Abstract_BB_Problem — Type

Abstract_BB_Problem #abstract branch and bound problem

Sould implement the methods below

Dionysos.Control.QuadraticStateControlFunction — Type

QuadraticStateControlFunction{T, MT<:AbstractMatrix{T}}

Quadratic function on state and input defined as x'Qx + u'Ru + 2x'Nu + 2x'q + 2u'r + v